+++ src/main/java/com/fleetmanager/vehiclefleetmanagement/dto/TripDetailsDTO.java
@@ -1,7 +1,6 @@
-package com.fleetmanager.vehiclefleetmanagement.DTO.trip;
+package com.fleetmanager.vehiclefleetmanagement.dto;
 
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -11,8 +10,11 @@ import java.util.UUID;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class TripSummaryDTO {
+public class TripDetailsDTO {
     private UUID id;
+    private UUID driverId;
+    private UUID vehicleId;
+    private String purpose;
     private Date startTime;
     private Date endTime;
     private String startLocation;
