+++ src/main/java/com/fleetmanager/vehiclefleetmanagement/dto/TripSummaryDTO.java
@@ -1,7 +1,6 @@
-package com.fleetmanager.vehiclefleetmanagement.DTO.trip;
+package com.fleetmanager.vehiclefleetmanagement.dto;
 
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -11,11 +10,8 @@ import java.util.UUID;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class TripDetailsDTO {
+public class TripSummaryDTO {
     private UUID id;
-    private UUID driverId;
-    private UUID vehicleId;
-    String purpose;
     private Date startTime;
     private Date endTime;
     private String startLocation;
