+++ src/main/java/com/fleetmanager/vehiclefleetmanagement/mapper/TripMapper.java
@@ -1,14 +1,12 @@
 package com.fleetmanager.vehiclefleetmanagement.mapper;
 
-import com.fleetmanager.vehiclefleetmanagement.DTO.trip.TripDetailsDTO;
-import com.fleetmanager.vehiclefleetmanagement.DTO.trip.TripSummaryDTO;
+import com.fleetmanager.vehiclefleetmanagement.dto.TripDetailsDTO;
+import com.fleetmanager.vehiclefleetmanagement.dto.TripSummaryDTO;
 import com.fleetmanager.vehiclefleetmanagement.entity.Trip;
 import org.mapstruct.BeanMapping;
 import org.mapstruct.Mapper;
 import org.mapstruct.MappingTarget;
 import org.mapstruct.NullValuePropertyMappingStrategy;
-
-import java.util.Optional;
 
 @Mapper(componentModel = "spring")
 public interface TripMapper {
